Fixed loading scene loading test 1
