+++ MonoEight.Template/Game/TestScene.cs
@@ -1,8 +1,4 @@
-using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Graphics;
 using MonoEight.Core;
-using MonoEight.Core.UserInput;
-using MonoEight.Core.UI;
 using MonoEight.Core.Scenes;
 
 namespace Template;
@@ -11,28 +7,8 @@ public class TestScene : Scene
 {
     protected override void Initialize()
     {
-        Camera.BackgroundColor = Color.LightYellow;
-    }
-
-    protected override void Update()
-    {
-        if (Input.IsPressed("A"))
-            SceneManager.Load("Test 2");
-
-        if (Input.IsPressed("Exit"))
-            SceneManager.Load("Loading");
-    }
-
-    protected override void Draw(SpriteBatch spriteBatch)
-    {
-        Canvas.DrawText(spriteBatch, "Test", FontSize.S, new(0, -20), MEColors.Red);
-        Canvas.DrawText(spriteBatch, "Test", FontSize.M, new(0, 0), MEColors.Red);
-        Canvas.DrawText(spriteBatch, "Lorem Ipsum is", FontSize.S, new(0, 10), MEColors.Red);
-        Canvas.DrawText(spriteBatch, "Simply dummy text.", FontSize.M, new(0, 20), MEColors.Red);
-
-        Debugger.DrawPixel(spriteBatch, new(0, 0), MEColors.Black);
-        Debugger.DrawPixel(spriteBatch, new(0, 5), MEColors.Blue);
-        Debugger.DrawPixel(spriteBatch, new(0, -5), MEColors.Green);
-        Debugger.DrawPixel(spriteBatch, new(0, -3), MEColors.Orange);
+        Camera.BackgroundColor = MEColors.Blue;
+        Add(new NewPlayer("PlayerTest", -16));
+        Add(new NewPlayer2("PlayerTest", 16));
     }
 }
