+++ MonoEight.Template/MonoEight/Scenes/LoadingScene.cs
@@ -40,6 +40,6 @@ public class LoadingScene : Scene
         _timer += Time.DeltaTime;
 
         if (_timer >= WAIT_TIME)
-            SceneManager.Load("Test 1");
+            SceneManager.Load("MainMenu");
     }
 }
